+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.4
+// swift-tools-version:5.5
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 //
 //  Created by Jesse Squires
@@ -29,8 +29,7 @@ let package = Package(
     targets: [
         .target(
             name: "AboutThisApp",
-            path: "Sources",
-            exclude: ["Info.plist"]),
+            path: "Sources"),
         .testTarget(name: "AboutThisAppTests",
                     dependencies: ["AboutThisApp"],
                     path: "Tests",